Add files via upload
File Naming Key/Matrix:-
W02 = Week 2
L03S = Lab 03 Submission

Additional Clarity Note:-
This file was uploaded within the due date, for verifications purposes the exact same pdf file (consisting of the flowcharts) were attached to Google Classroom when the submission for the github link was made as well. As to why is this file being reuploaded, when i was previously uploaded on github. Earlier, it was in the repository but at that time there were no folders in the repository. Hence, to organize my work and for the ease of navigation, I shifted to an approach where I store every weeks file in distinct folders. Because I moved this file into Week2's folder (which we are currently in) the pdf had loading errors. Therefore, it was necessary that the file was reuploaded to enssure it is found by the checker at the time of marking. The uploading of this file earlier within the due date on github can also be verified via github's editing history when we commit changes, witth their respective due dates when that particular change/edit was committed.